test: BoxOption
